Move all the variables into the hosts file
The change pushes all the variables outside the task files into the
hosts_vars file. The final idea is to have the hosts_vars file act as
the sole interface with the user (under the assumption that new features
don't get introduced). This means that all transient changes happen in
this hosts_vars file.