Remove the ansible user setup
Since most of the times the user should already be set in the
installation process (for both Arch and Ubuntu), the user setup has no
effect and only adds clutter to the process. The only useful thing left
of the user process is making sure the correct home / shell have been
set. Everything else (including the user password, groups and so on)
should already be taken care of in the manual installation process.